move `HomeNav.vue` content to `Footer.vue`
Currently there is no need to have the `HomeNav` content nested within
the footer, so it was moved to the `Footer` component. This is helpful
so that we don't need to layer the emitted events.
